# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouBike 2.0 Taipei Station API",
    "description": "\n以 **臺北市 YouBike 2.0 站點即時資訊** Open Data 為主題的 RESTful API。\n\n資料來源：臺北市政府交通局（data.taipei）\n`https://tcgbusfs.blob.core.windows.net/dotapp/youbike/v2/youbike_immediate.json`\n\n## Resources\n| Resource | 說明 |\n|---|---|\n| **Areas** | 行政區（由站點資料的 `sarea` 欄位整理而成） |\n| **Stations** | YouBike 2.0 站點：位置、車位、即時可借 / 可還數量 |\n| **Stats & Admin** | 車況統計與 Open Data 重新匯入 |\n\n所有端點皆以 `/api/v1` 為前綴。資料儲存於 SQLite，首次啟動若資料庫為空會自動匯入 `data/youbike_stations.json`。\n",
    "contact": {
      "name": "IoTApp Open Data RESTful API Homework"
    },
    "license": {
      "name": "政府資料開放授權條款－第1版",
      "url": "https://data.gov.tw/license"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/areas": {
      "get": {
        "tags": [
          "Areas"
        ],
        "summary": "列出所有行政區",
        "operationId": "list_areas_api_v1_areas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AreaOut"
                  },
                  "type": "array",
                  "title": "Response List Areas Api V1 Areas Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Areas"
        ],
        "summary": "新增行政區",
        "operationId": "create_area_api_v1_areas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AreaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaOut"
                }
              }
            }
          },
          "409": {
            "description": "名稱重複",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/areas/{area_id}": {
      "get": {
        "tags": [
          "Areas"
        ],
        "summary": "取得單一行政區",
        "operationId": "get_area_api_v1_areas__area_id__get",
        "parameters": [
          {
            "name": "area_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Area Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Areas"
        ],
        "summary": "更新行政區",
        "operationId": "update_area_api_v1_areas__area_id__put",
        "parameters": [
          {
            "name": "area_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Area Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AreaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Areas"
        ],
        "summary": "刪除行政區",
        "operationId": "delete_area_api_v1_areas__area_id__delete",
        "parameters": [
          {
            "name": "area_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Area Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/areas/{area_id}/stations": {
      "get": {
        "tags": [
          "Areas"
        ],
        "summary": "列出行政區內的站點",
        "operationId": "list_area_stations_api_v1_areas__area_id__stations_get",
        "parameters": [
          {
            "name": "area_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Area Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StationOut_"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AreaCreate": {
        "properties": {
          "name_zh": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name Zh",
            "examples": [
              "大安區"
            ]
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En",
            "examples": [
              "Daan Dist."
            ]
          }
        },
        "type": "object",
        "required": [
          "name_zh"
        ],
        "title": "AreaCreate"
      },
      "AreaOut": {
        "properties": {
          "name_zh": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name Zh",
            "examples": [
              "大安區"
            ]
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En",
            "examples": [
              "Daan Dist."
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "station_count": {
            "type": "integer",
            "title": "Station Count",
            "description": "該行政區站點數",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name_zh",
          "id"
        ],
        "title": "AreaOut"
      },
...
      "AreaUpdate": {
        "properties": {
          "name_zh": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Zh"
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En"
          }
        },
        "type": "object",
        "title": "AreaUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "Message"
      },
      "Page_StationOut_": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "符合條件的總筆數"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StationOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "Page[StationOut]"
      },
...
      "StationOut": {
        "properties": {
          "name_zh": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name Zh",
            "examples": [
              "YouBike2.0_捷運科技大樓站"
            ]
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En",
            "examples": [
              "YouBike2.0_MRT Technology Bldg. Sta."
            ]
          },
          "area_id": {
            "type": "integer",
            "title": "Area Id",
            "description": "所屬行政區 id（參考 /areas）",
            "examples": [
              1
            ]
          },
          "address_zh": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Zh",
            "examples": [
              "復興南路二段235號前"
            ]
          },
          "address_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Address En",
            "examples": [
              "No.235， Sec. 2， Fuxing S. Rd."
            ]
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "examples": [
              25.02605
            ]
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "examples": [
              121.5436
            ]
          },
          "total_docks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Docks",
            "description": "總車位數",
            "examples": [
              28
            ]
          },
          "available_bikes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Available Bikes",
            "description": "可借車輛數",
            "examples": [
              5
            ]
          },
          "available_docks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Available Docks",
            "description": "可還空位數",
            "examples": [
              22
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "站點是否營運中",
            "default": true
          },
          "info_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info Time",
            "description": "站點資料回報時間"
          },
          "sno": {
            "type": "string",
            "title": "Sno"
          },
          "area_name": {
            "type": "string",
            "title": "Area Name",
            "description": "行政區中文名稱"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name_zh",
          "area_id",
          "latitude",
          "longitude",
          "total_docks",
          "available_bikes",
          "available_docks",
          "sno",
          "area_name",
          "updated_at"
        ],
        "title": "StationOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}